celeryBackend/celery_requirements.txt
celery
datasets
asyncpg
flower
loguru
pydantic
python-dotenv
redis
torch
torchaudio==2.6.0
torchvision==0.21.0
tornado==6.4.2
transformers==4.50.1
tsidpy==1.1.5
typer==0.15.2
typing_extensions
yarl==1.19.0
watchfiles==1.0.4
psycopg2